requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
streamlit==1.22.0
python-multipart==0.0.5
requests==2.28.2
pandas==1.5.3
pillow==9.4.0
opencv-python==4.7.0.72
numpy==1.23.5
deepface==0.0.79
pydantic==1.10.7 